+++ frontend/.prettierignore
@@ -1,3 +1,3 @@
 node_modules/
 .next/
-frontend/public/*.js
+public/*.js
