--- frontend/src/components/Skills.tsx
@@ -1,5 +1,0 @@
-const Skills = () => {
-    return <div></div>;
-};
-
-export default Skills;
